refactor(parquet): split `arrow_reader/selection` into smaller modules
`selection/mod.rs` had grown past 2100 lines and mixed the `RowSelection`
type, the set algebra, the page range mapping and the cursor machinery.

Split it by concern, with one module per selection backing:

* `mod.rs`      - `RowSelection` and its public API, dispatching on backing
* `selector.rs` - run length backing: `RowSelector` and its primitives
* `boolean.rs`  - bitmap backing: `MaskSelection` and its primitives
* `algebra.rs`  - `and_then`, `intersection` and `union`
* `ranges.rs`   - page byte ranges and batch boundary expansion
* `cursor.rs`   - `RowSelectionCursor` and the policy / strategy types

No functional change: bodies are moved verbatim and the public API is
unchanged.
