+++ parquet/src/arrow/arrow_reader/selection/boolean.rs
@@ -15,22 +15,31 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use super::{LoadedRowRanges, RowSelection, RowSelectionInner, RowSelector};
-use crate::errors::ParquetError;
-use arrow_array::BooleanArray;
+//! The bitmap backed representation of a [`RowSelection`] and the primitives
+//! operating on it: conversion to and from the run length ([`RowSelector`])
+//! form, and the transforms backing `split_off`, `trim`, `offset` and `limit`.
+//!
+//! The bitwise set algebra lives in the `algebra` module.
+//!
+//! [`RowSelection`]: crate::arrow::arrow_reader::RowSelection
+
+use super::RowSelector;
 use arrow_buffer::bit_iterator::BitSliceIterator;
 use arrow_buffer::{BooleanBuffer, BooleanBufferBuilder, Buffer};
-use std::cmp::Ordering;
-use std::sync::{Arc, OnceLock};
+use std::sync::OnceLock;
 
 /// Mask-backed [`RowSelection`] storage.
 ///
-/// `selectors` is only populated if callers use the borrowed [`RowSelection::iter`]
-/// compatibility API. Internal paths that can stream or consume the bitmap avoid
-/// this cache.
+/// `selectors` is only populated if callers use the borrowed
+/// [`RowSelection::iter`] compatibility API. Internal paths that can stream or
+/// consume the bitmap avoid this cache.
 ///
-/// `count` caches the popcount; `RowSelection::split_off` propagates it to
+/// `count` caches the popcount; [`RowSelection::split_off`] propagates it to
 /// both halves so repeated `row_count()` calls do not rescan the bitmap.
+///
+/// [`RowSelection`]: crate::arrow::arrow_reader::RowSelection
+/// [`RowSelection::iter`]: crate::arrow::arrow_reader::RowSelection::iter
+/// [`RowSelection::split_off`]: crate::arrow::arrow_reader::RowSelection::split_off
 #[derive(Debug)]
 pub(crate) struct MaskSelection {
     mask: BooleanBuffer,
@@ -107,6 +116,9 @@ impl Clone for MaskSelection {
 ///     for run in MaskRunIter::new(mask) { ... }
 /// }
 /// ```
+///
+/// [`RowSelection::iter`]: crate::arrow::arrow_reader::RowSelection::iter
+/// [`RowSelection::as_mask`]: crate::arrow::arrow_reader::RowSelection::as_mask
 #[derive(Debug)]
 pub struct MaskRunIter<'a> {
     slices: BitSliceIterator<'a>,
@@ -166,178 +178,6 @@ impl Iterator for MaskRunIter<'_> {
     }
 }
 
-/// Cursor for iterating a mask-backed [`RowSelection`]
-///
-/// This is best for dense selections where there are many small skips
-/// or selections. For example, selecting every other row.
-///
-/// When page pruning produces sparse column data, `loaded_row_ranges` limits
-/// each decoded chunk to rows whose pages are loaded for every projected leaf.
-/// For example, two projected columns can have different page boundaries:
-///
-/// ```text
-/// Row ranges:       [0, 4) [4, 6) [6, 8) [8, 10) [10, 12)
-/// Selection mask:   1000   00     00     00      01
-/// Column A pages:   loaded | missing [4, 8) | loaded [8, 12)
-/// Column B pages:   loaded [0, 6) | missing [6, 10) | loaded
-/// LoadedRowRanges:  [0, 4)                         [10, 12)
-/// ```
-///
-/// The first chunk decodes `[0, 4)` with mask `1000`. The next chunk skips to
-/// row 11 and decodes `[11, 12)` with mask `1`. The loaded ranges are decode
-/// boundaries, not output batch boundaries: [`ParquetRecordBatchReader`]
-/// accumulates both chunks and applies the combined mask `10001` once.
-///
-/// [`ParquetRecordBatchReader`]: crate::arrow::arrow_reader::ParquetRecordBatchReader
-#[derive(Debug)]
-pub struct MaskCursor {
-    pub(super) mask: BooleanBuffer,
-    /// Current absolute offset into the selection
-    pub(super) position: usize,
-    /// Row ranges whose backing pages are loaded for every projected column.
-    pub(super) loaded_row_ranges: Option<Arc<LoadedRowRanges>>,
-}
-
-impl MaskCursor {
-    /// Returns `true` when no further rows remain
-    pub fn is_empty(&self) -> bool {
-        self.position >= self.mask.len()
-    }
-
-    /// Advance through the mask representation, producing the next chunk summary
-    pub fn next_mask_chunk(&mut self, batch_size: usize) -> Option<MaskChunk> {
-        if self.is_empty() {
-            return None;
-        }
-
-        Some(self.next_mask_chunk_non_empty(batch_size))
-    }
-
-    /// Produces the next chunk for a non-empty, trailing-skip-free mask.
-    fn next_mask_chunk_non_empty(&mut self, batch_size: usize) -> MaskChunk {
-        debug_assert!(!self.is_empty());
-
-        let (initial_skip, chunk_rows, selected_rows, mask_start, end_position) = {
-            let mask = &self.mask;
-            let start_position = self.position;
-            let mut cursor = start_position;
-            let mut initial_skip = 0;
-
-            while cursor < mask.len() && !mask.value(cursor) {
-                initial_skip += 1;
-                cursor += 1;
-            }
-            debug_assert!(
-                cursor < mask.len(),
-                "ReadPlan must remove trailing skips from Mask selections"
-            );
-
-            let mask_start = cursor;
-            let mut chunk_rows = 0;
-            let mut selected_rows = 0;
-
-            // Advance until enough rows have been selected to satisfy the batch size,
-            // or until the mask is exhausted. This mirrors the behaviour of the legacy
-            // `RowSelector` queue-based iteration.
-            while cursor < mask.len() && selected_rows < batch_size {
-                chunk_rows += 1;
-                if mask.value(cursor) {
-                    selected_rows += 1;
-                }
-                cursor += 1;
-            }
-
-            (initial_skip, chunk_rows, selected_rows, mask_start, cursor)
-        };
-
-        self.position = end_position;
-
-        MaskChunk {
-            initial_skip,
-            chunk_rows,
-            selected_rows,
-            mask_start,
-        }
-    }
-
-    /// Returns the next non-empty mask chunk without crossing an unloaded row range.
-    ///
-    /// The [`ReadPlan`](crate::arrow::arrow_reader::ReadPlan) removes trailing
-    /// skips before constructing this cursor. Callers therefore only invoke
-    /// this method for a non-empty mask that has another selected row.
-    pub(crate) fn next_chunk(&mut self, batch_size: usize) -> Result<MaskChunk, ParquetError> {
-        debug_assert!(batch_size > 0);
-        debug_assert!(!self.is_empty());
-
-        if self.loaded_row_ranges.is_none() {
-            return Ok(self.next_mask_chunk_non_empty(batch_size));
-        }
-
-        let start_position = self.position;
-        let mut cursor = start_position;
-        while cursor < self.mask.len() && !self.mask.value(cursor) {
-            cursor += 1;
-        }
-
-        debug_assert!(
-            cursor < self.mask.len(),
-            "ReadPlan must remove trailing skips from Mask selections"
-        );
-
-        let loaded_range_end = self
-            .loaded_row_ranges
-            .as_ref()
-            .and_then(|ranges| ranges.end_containing(cursor))
-            .ok_or_else(|| {
-                ParquetError::General(format!(
-                    "Internal Error: selected row {cursor} has no loaded page range"
-                ))
-            })?;
-
-        let mask_start = cursor;
-        let mut selected_rows = 0;
-        while cursor < loaded_range_end && cursor < self.mask.len() && selected_rows < batch_size {
-            if self.mask.value(cursor) {
-                selected_rows += 1;
-            }
-            cursor += 1;
-        }
-
-        self.position = cursor;
-        Ok(MaskChunk {
-            initial_skip: mask_start - start_position,
-            chunk_rows: cursor - mask_start,
-            selected_rows,
-            mask_start,
-        })
-    }
-
-    /// Materialise the boolean values for a mask-backed chunk
-    pub fn mask_values_for(&self, chunk: &MaskChunk) -> Result<BooleanArray, ParquetError> {
-        if chunk.mask_start.saturating_add(chunk.chunk_rows) > self.mask.len() {
-            return Err(ParquetError::General(
-                "Internal Error: MaskChunk exceeds mask length".to_string(),
-            ));
-        }
-        Ok(BooleanArray::from(
-            self.mask.slice(chunk.mask_start, chunk.chunk_rows),
-        ))
-    }
-}
-
-/// Result of computing the next chunk to read when using a [`MaskCursor`]
-#[derive(Debug)]
-pub struct MaskChunk {
-    /// Number of leading rows to skip before reaching selected rows
-    pub initial_skip: usize,
-    /// Total rows covered by this chunk (selected + skipped)
-    pub chunk_rows: usize,
-    /// Rows actually selected within the chunk
-    pub selected_rows: usize,
-    /// Starting offset within the mask where the chunk begins
-    pub mask_start: usize,
-}
-
 /// Materialize a [`BooleanBuffer`] into its RLE form.
 pub(crate) fn mask_to_selectors(mask: &BooleanBuffer) -> Vec<RowSelector> {
     let total_rows = mask.len();
@@ -384,133 +224,6 @@ pub(super) fn mask_has_at_least_runs(mask: &BooleanBuffer, min_runs: usize) -> b
     }
 
     run_count + usize::from(last_end < total_rows) >= min_runs
-}
-
-/// Bitwise AND of two mask-backed selections. Longer side's tail passes through.
-pub(super) fn intersect_masks(l: &BooleanBuffer, r: &BooleanBuffer) -> BooleanBuffer {
-    if l.len() == r.len() {
-        return l & r;
-    }
-    let common = l.len().min(r.len());
-    let head = &l.slice(0, common) & &r.slice(0, common);
-    let (longer, longer_len) = if l.len() > r.len() {
-        (l, l.len())
-    } else {
-        (r, r.len())
-    };
-    let tail = longer.slice(common, longer_len - common);
-    let mut builder = BooleanBufferBuilder::new(longer_len);
-    builder.append_buffer(&head);
-    builder.append_buffer(&tail);
-    builder.finish()
-}
-
-/// Bitwise OR of two mask-backed selections. Longer side's tail passes through.
-pub(super) fn union_masks(l: &BooleanBuffer, r: &BooleanBuffer) -> BooleanBuffer {
-    if l.len() == r.len() {
-        return l | r;
-    }
-    let common = l.len().min(r.len());
-    let head = &l.slice(0, common) | &r.slice(0, common);
-    let (longer, longer_len) = if l.len() > r.len() {
-        (l, l.len())
-    } else {
-        (r, r.len())
-    };
-    let tail = longer.slice(common, longer_len - common);
-    let mut builder = BooleanBufferBuilder::new(longer_len);
-    builder.append_buffer(&head);
-    builder.append_buffer(&tail);
-    builder.finish()
-}
-
-/// Applies `other` to the selected rows of `mask`, preserving the original row domain.
-pub(super) fn and_then_mask(mask: &BooleanBuffer, other: &RowSelection) -> BooleanBuffer {
-    match &other.inner {
-        RowSelectionInner::Mask(other_mask) => and_then_masks(mask, other_mask.mask()),
-        RowSelectionInner::Selectors(selectors) => {
-            and_then_mask_from_selectors(mask, selectors.iter().copied())
-        }
-    }
-}
-
-fn and_then_mask_from_selectors<I>(mask: &BooleanBuffer, other: I) -> BooleanBuffer
-where
-    I: IntoIterator<Item = RowSelector>,
-{
-    let mut builder = BooleanBufferBuilder::new(mask.len());
-    let mut other_iter = other.into_iter();
-    let mut current = other_iter.next();
-    let mut cursor = 0usize;
-
-    // Iterate only over the set positions in `mask`; the gaps of unset bits
-    // are filled in bulk with `append_n` instead of bit-by-bit.
-    for set_idx in mask.set_indices() {
-        if set_idx > cursor {
-            builder.append_n(set_idx - cursor, false);
-        }
-        cursor = set_idx + 1;
-
-        while current.as_ref().is_some_and(|s| s.row_count == 0) {
-            current = other_iter.next();
-        }
-        let selector = current
-            .as_mut()
-            .expect("selection contains less than the number of selected rows");
-        let selected = !selector.skip;
-        selector.row_count -= 1;
-        builder.append(selected);
-    }
-    if cursor < mask.len() {
-        builder.append_n(mask.len() - cursor, false);
-    }
-
-    if current.is_some_and(|s| s.row_count != 0) || other_iter.any(|s| s.row_count != 0) {
-        panic!("selection exceeds the number of selected rows");
-    }
-
-    builder.finish()
-}
-
-fn and_then_masks(mask: &BooleanBuffer, other: &BooleanBuffer) -> BooleanBuffer {
-    let selected_count = mask.count_set_bits();
-    match other.len().cmp(&selected_count) {
-        Ordering::Less => panic!("selection contains less than the number of selected rows"),
-        Ordering::Greater => panic!("selection exceeds the number of selected rows"),
-        Ordering::Equal => {}
-    }
-
-    let other_true_count = other.count_set_bits();
-    if other_true_count == 0 {
-        return BooleanBuffer::new_unset(mask.len());
-    }
-    if other_true_count == selected_count {
-        return mask.clone();
-    }
-
-    let mut builder = BooleanBufferBuilder::new(mask.len());
-    let mut outer_set_indices = mask.set_indices();
-    let mut next_selected_ordinal = 0usize;
-    let mut cursor = 0usize;
-
-    for selected_ordinal in other.set_indices() {
-        let skip = selected_ordinal - next_selected_ordinal;
-        let set_idx = outer_set_indices
-            .nth(skip)
-            .expect("validated other length matches selected row count");
-        if set_idx > cursor {
-            builder.append_n(set_idx - cursor, false);
-        }
-        builder.append(true);
-        cursor = set_idx + 1;
-        next_selected_ordinal = selected_ordinal + 1;
-    }
-
-    if cursor < mask.len() {
-        builder.append_n(mask.len() - cursor, false);
-    }
-
-    builder.finish()
 }
 
 /// Split a mask into `(head, tail)` at `row_count`, preserving an empty mask tail
@@ -624,6 +337,7 @@ pub(super) fn boolean_mask_from_selectors(selectors: &[RowSelector]) -> BooleanB
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::arrow::arrow_reader::selection::{RowSelection, RowSelectionInner};
     use arrow_array::BooleanArray;
     use rand::{Rng, rng};
 
@@ -824,72 +538,6 @@ mod tests {
     }
 
     #[test]
-    fn test_mask_and_then_preserves_backing() {
-        let outer_bits = vec![false, true, true, false, true, false, true];
-        let inner_bits = vec![true, false, true, false];
-        let outer_mask = RowSelection::from_boolean_buffer(BooleanBuffer::from(outer_bits.clone()));
-        let inner = RowSelection::from_filters(&[BooleanArray::from(inner_bits.clone())]);
-
-        let result = outer_mask.and_then(&inner);
-        assert!(result.as_mask().is_some());
-
-        let outer_selectors = RowSelection::from_filters(&[BooleanArray::from(outer_bits)]);
-        let expected = outer_selectors.and_then(&inner);
-        assert_eq!(result, expected);
-
-        let result_mask = result.as_mask().unwrap();
-        let actual_bits: Vec<_> = (0..result_mask.len())
-            .map(|i| result_mask.value(i))
-            .collect();
-        assert_eq!(
-            actual_bits,
-            vec![false, true, false, false, true, false, false]
-        );
-    }
-
-    #[test]
-    fn test_mask_and_then_mask_preserves_backing() {
-        let outer_bits = vec![false, true, true, false, true, false, true, true];
-        let inner_bits = vec![false, true, false, true, false];
-        let outer_mask = RowSelection::from_boolean_buffer(BooleanBuffer::from(outer_bits.clone()));
-        let inner_mask = RowSelection::from_boolean_buffer(BooleanBuffer::from(inner_bits));
-
-        let result = outer_mask.and_then(&inner_mask);
-        assert!(result.as_mask().is_some());
-
-        let outer_selectors = RowSelection::from_filters(&[BooleanArray::from(outer_bits)]);
-        let inner_selectors = RowSelection::from_filters(&[BooleanArray::from(vec![
-            false, true, false, true, false,
-        ])]);
-        assert_eq!(result, outer_selectors.and_then(&inner_selectors));
-
-        let result_mask = result.as_mask().unwrap();
-        let actual_bits: Vec<_> = (0..result_mask.len())
-            .map(|i| result_mask.value(i))
-            .collect();
-        assert_eq!(
-            actual_bits,
-            vec![false, false, true, false, false, false, true, false]
-        );
-    }
-
-    #[test]
-    fn test_selector_and_then_mask() {
-        let outer =
-            RowSelection::from_filters(&[BooleanArray::from(vec![false, true, true, false, true])]);
-        let inner = RowSelection::from_boolean_buffer(BooleanBuffer::from(vec![true, false, true]));
-
-        let result = outer.and_then(&inner);
-        assert!(result.as_mask().is_none());
-        assert_eq!(
-            result,
-            RowSelection::from_filters(&[BooleanArray::from(vec![
-                false, true, false, false, true,
-            ])])
-        );
-    }
-
-    #[test]
     fn test_mask_offset_past_end_preserves_empty_mask_backing() {
         let selection =
             RowSelection::from_boolean_buffer(BooleanBuffer::from(vec![true, false, true]))
@@ -912,161 +560,6 @@ mod tests {
         assert_eq!(mask.len(), 4);
         let actual_bits: Vec<_> = (0..mask.len()).map(|i| mask.value(i)).collect();
         assert_eq!(actual_bits, vec![false, true, false, true]);
-    }
-
-    #[test]
-    fn test_mask_intersection_uses_bitwise() {
-        let a_bits = vec![true, true, false, true, false, true];
-        let b_bits = vec![true, false, true, true, true, false];
-        let a = RowSelection::from_boolean_buffer(BooleanBuffer::from(a_bits.clone()));
-        let b = RowSelection::from_boolean_buffer(BooleanBuffer::from(b_bits.clone()));
-
-        let r = a.intersection(&b);
-        assert!(r.as_mask().is_some());
-
-        let expected: Vec<bool> = a_bits.iter().zip(&b_bits).map(|(x, y)| *x && *y).collect();
-        let expected_sel = RowSelection::from_filters(&[BooleanArray::from(expected)]);
-        assert_eq!(r, expected_sel);
-    }
-
-    #[test]
-    fn test_mask_union_uses_bitwise() {
-        let a_bits = vec![true, false, false, true, false, false];
-        let b_bits = vec![false, true, false, false, true, false];
-        let a = RowSelection::from_boolean_buffer(BooleanBuffer::from(a_bits.clone()));
-        let b = RowSelection::from_boolean_buffer(BooleanBuffer::from(b_bits.clone()));
-
-        let r = a.union(&b);
-        assert!(r.as_mask().is_some());
-
-        let expected: Vec<bool> = a_bits.iter().zip(&b_bits).map(|(x, y)| *x || *y).collect();
-        let expected_sel = RowSelection::from_filters(&[BooleanArray::from(expected)]);
-        assert_eq!(r, expected_sel);
-    }
-
-    #[test]
-    fn test_mixed_mask_selector_intersection_and_union() {
-        let mask_bits = vec![true, false, true, false, true, false];
-        let selector_bits = vec![false, true, true, false, false, true];
-        let mask = RowSelection::from_boolean_buffer(BooleanBuffer::from(mask_bits.clone()));
-        let selectors = RowSelection::from_filters(&[BooleanArray::from(selector_bits.clone())]);
-
-        let intersection_bits: Vec<_> = mask_bits
-            .iter()
-            .zip(&selector_bits)
-            .map(|(x, y)| *x && *y)
-            .collect();
-        let expected_intersection =
-            RowSelection::from_filters(&[BooleanArray::from(intersection_bits)]);
-        assert_eq!(mask.intersection(&selectors), expected_intersection);
-        assert_eq!(selectors.intersection(&mask), expected_intersection);
-
-        let union_bits: Vec<_> = mask_bits
-            .iter()
-            .zip(&selector_bits)
-            .map(|(x, y)| *x || *y)
-            .collect();
-        let expected_union = RowSelection::from_filters(&[BooleanArray::from(union_bits)]);
-        assert_eq!(mask.union(&selectors), expected_union);
-        assert_eq!(selectors.union(&mask), expected_union);
-    }
-
-    #[test]
-    fn test_mask_intersection_uneven_passes_tail_through() {
-        let a_bits = vec![true, true, true, true, true];
-        let b_bits = vec![true, false, true];
-        let a = RowSelection::from_boolean_buffer(BooleanBuffer::from(a_bits));
-        let b = RowSelection::from_boolean_buffer(BooleanBuffer::from(b_bits));
-
-        let r = a.intersection(&b);
-        let r_mask = r.as_mask().unwrap();
-        assert_eq!(r_mask.len(), 5);
-        let bits: Vec<bool> = (0..5).map(|i| r_mask.value(i)).collect();
-        assert_eq!(bits, vec![true, false, true, true, true]);
-
-        // Swapped operands: the right side is longer and its tail passes through.
-        let a = RowSelection::from_boolean_buffer(BooleanBuffer::from(vec![true, false, true]));
-        let b = RowSelection::from_boolean_buffer(BooleanBuffer::from(vec![
-            true, true, true, false, true,
-        ]));
-        let r = a.intersection(&b);
-        let r_mask = r.as_mask().unwrap();
-        assert_eq!(r_mask.len(), 5);
-        let bits: Vec<bool> = (0..5).map(|i| r_mask.value(i)).collect();
-        assert_eq!(bits, vec![true, false, true, false, true]);
-    }
-
-    #[test]
-    fn test_mask_and_then_none_selected_returns_all_unset() {
-        let outer = RowSelection::from_boolean_buffer(BooleanBuffer::from(vec![
-            false, true, true, false, true,
-        ]));
-        let inner =
-            RowSelection::from_boolean_buffer(BooleanBuffer::from(vec![false, false, false]));
-
-        let result = outer.and_then(&inner);
-        let mask = result.as_mask().unwrap();
-        assert_eq!(mask.len(), 5);
-        assert_eq!(mask.count_set_bits(), 0);
-    }
-
-    #[test]
-    fn test_mixed_backing_equality_mismatches() {
-        let mask =
-            RowSelection::from_boolean_buffer(BooleanBuffer::from(vec![true, false, true, true]));
-
-        // Total row counts differ
-        let longer = RowSelection::from(vec![
-            RowSelector::select(1),
-            RowSelector::skip(1),
-            RowSelector::select(2),
-            RowSelector::skip(1),
-        ]);
-        assert_ne!(mask, longer);
-        assert_ne!(longer, mask);
-
-        // A selected bit falls inside a skip run
-        let skip_overlap = RowSelection::from(vec![RowSelector::skip(2), RowSelector::select(2)]);
-        assert_ne!(mask, skip_overlap);
-
-        // Select run boundaries do not line up
-        let misaligned = RowSelection::from(vec![
-            RowSelector::select(2),
-            RowSelector::skip(1),
-            RowSelector::select(1),
-        ]);
-        assert_ne!(mask, misaligned);
-
-        let equal = RowSelection::from(vec![
-            RowSelector::select(1),
-            RowSelector::skip(1),
-            RowSelector::select(2),
-        ]);
-        assert_eq!(mask, equal);
-        assert_eq!(equal, mask);
-    }
-
-    #[test]
-    fn test_mask_union_uneven_passes_tail_through() {
-        let a_bits = vec![true, false, true];
-        let b_bits = vec![false, true, false, true, false];
-        let a = RowSelection::from_boolean_buffer(BooleanBuffer::from(a_bits));
-        let b = RowSelection::from_boolean_buffer(BooleanBuffer::from(b_bits));
-
-        let r = a.union(&b);
-        let r_mask = r.as_mask().unwrap();
-        assert_eq!(r_mask.len(), 5);
-        let bits: Vec<bool> = (0..5).map(|i| r_mask.value(i)).collect();
-        assert_eq!(bits, vec![true, true, true, true, false]);
-
-        let a = RowSelection::from_boolean_buffer(BooleanBuffer::from(vec![
-            false, true, false, false, true,
-        ]));
-        let b = RowSelection::from_boolean_buffer(BooleanBuffer::from(vec![true, false, false]));
-        let r = a.union(&b);
-        let r_mask = r.as_mask().unwrap();
-        let bits: Vec<bool> = (0..5).map(|i| r_mask.value(i)).collect();
-        assert_eq!(bits, vec![true, true, false, false, true]);
     }
 
     #[test]
@@ -1130,61 +623,6 @@ mod tests {
         let trimmed = s.trim();
         assert!(trimmed.as_mask().is_some());
         assert_eq!(trimmed.as_mask().unwrap().len(), 0);
-    }
-
-    #[test]
-    fn test_from_iter_all_mask_preserves_mask_backing() {
-        let a_bits = vec![true, false, true, true];
-        let b_bits = vec![false, true, false];
-        let c_bits = vec![true, true, false, false, true];
-
-        let parts = vec![
-            RowSelection::from_boolean_buffer(BooleanBuffer::from(a_bits.clone())),
-            RowSelection::from_boolean_buffer(BooleanBuffer::from(b_bits.clone())),
-            RowSelection::from_boolean_buffer(BooleanBuffer::from(c_bits.clone())),
-        ];
-        let collected: RowSelection = parts.into_iter().collect();
-
-        let combined = a_bits
-            .iter()
-            .chain(b_bits.iter())
-            .chain(c_bits.iter())
-            .copied()
-            .collect::<Vec<_>>();
-        let expected = RowSelection::from_filters(&[BooleanArray::from(combined)]);
-
-        assert!(collected.as_mask().is_some());
-        assert_eq!(collected, expected);
-    }
-
-    #[test]
-    fn test_from_iter_mixed_backing_falls_back_to_selectors() {
-        let a_bits = vec![true, false, true];
-        let b_selectors = vec![RowSelector::skip(2), RowSelector::select(3)];
-        let c_bits = vec![false, true];
-
-        let parts = vec![
-            RowSelection::from_boolean_buffer(BooleanBuffer::from(a_bits.clone())),
-            RowSelection::from(b_selectors),
-            RowSelection::from_boolean_buffer(BooleanBuffer::from(c_bits.clone())),
-        ];
-        let collected: RowSelection = parts.into_iter().collect();
-
-        assert!(collected.as_mask().is_none());
-
-        let combined_bits = vec![
-            true, false, true, false, false, true, true, true, false, true,
-        ];
-        let expected = RowSelection::from_filters(&[BooleanArray::from(combined_bits)]);
-        assert_eq!(collected, expected);
-    }
-
-    #[test]
-    fn test_from_iter_empty_yields_empty_selection() {
-        let collected: RowSelection = std::iter::empty::<RowSelection>().collect();
-        assert_eq!(collected, RowSelection::default());
-        assert!(collected.as_mask().is_some());
-        assert_eq!(collected.as_mask().unwrap().len(), 0);
     }
 
     #[test]
